fixed linter issues
